+++ contracts/libraries/DataTypes.sol
@@ -6,8 +6,8 @@ library DataTypes {
     struct Block {
         bytes32 rootHash;
         bytes32 intentHash; // hash of L2-to-L1 commitment sync transitions
-        uint256 blockTime; // blockNum when this rollup block is committed
-        uint256 blockSize; // number of transitions in the block
+        uint128 blockTime; // blockNum when this rollup block is committed
+        uint128 blockSize; // number of transitions in the block
     }
 
     struct DepositTransition {
